refactorings: SideMenuBody.js 컴포넌트 파일 분리
# why
one component in one file
